implement query database , bind adapter MovieAdapter.kt, implement logic

+++ app/src/main/java/com/magdalena/movietoday/database/FavoriteMovie.kt
@@ -11,5 +11,5 @@ data class FavoriteMovie(
     val id: Int? = null,
 
     @ColumnInfo
-    val movieId: Int
+    val movieId: Long
 )
