+++ app/src/main/java/com/magdalena/movietoday/database/FavoriteMovieDao.kt
@@ -13,7 +13,7 @@ interface FavoriteMovieDao {
     fun saveFavoriteMovie(favoriteMovie: FavoriteMovie)
 
     @Query("DELETE FROM favorite_movie WHERE movieId = :movieId")
-    fun deleteFavoriteMovie(movieId: Int)
+    fun deleteFavoriteMovie(movieId: Long)
 
     @Query("SELECT * FROM favorite_movie")
     fun getFavoriteMovieIds(): MutableList<FavoriteMovie>
